Fix flipped image on write
Previous behavior: read/2 resulted in the pixmap list in the erl_image
record being sorted backwards by rownum, due to the use of the idiom of
prepending new elements to a list as the function is called recursively
(in the fun in read/2 itself).

write/2 would sort ascending by rownum, but that would then be reversed
via the fun given to lists:foldl, which also builds up the output list
by prepending.

The pixmap is thus output reversed.

New behavior: the pixmap list is kept in ascending order upon read (via
lists:reverse in raw_data/7). This is mostly on principle, since in any
case, write/2 now sorts *descending* before folding, resulting in the
output of the fold winding up being sorted in ascending order, as
desired.